moved declaration to .d.ts
